+++ R/predictEvent.R
@@ -142,9 +142,6 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
         df <- df %>% dplyr::mutate(
             treatment_description = paste0("Treatment ", .data$treatment))
       }
-
-      df$treatment_description = stats::reorder(
-        as.factor(df$treatment_description), df$treatment)
     }
 
     if (!is.null(newSubjects)) {
@@ -152,9 +149,6 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
       newSubjects <- newSubjects %>% dplyr::mutate(
           treatment_description = paste0("Treatment ", .data$treatment))
       }
-
-      newSubjects$treatment_description = stats::reorder(
-        as.factor(newSubjects$treatment_description), newSubjects$treatment)
     }
 
   } else {  # treat as a special case of by-treatment calculation
@@ -428,6 +422,7 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
         dplyr::summarise(nenrolled = sum(.data$arrivalTime <= .data$t) + n0,
                          .groups = "drop_last") %>%
         dplyr::summarise(n = quantile(.data$nenrolled, probs = 0.5),
+                         pilevel = pilevel,
                          lower = quantile(.data$nenrolled, probs = plower),
                          upper = quantile(.data$nenrolled, probs = pupper),
                          mean = mean(.data$nenrolled),
@@ -438,20 +433,22 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
 
     if (!is.null(df)) {
       # day 1
-      df0 <- dplyr::tibble(t = 1, n = 0, lower = NA, upper = NA,
-                           mean = 0, var = 0)
+      df0 <- dplyr::tibble(t = 1, n = 0, pilevel = pilevel,
+                           lower = NA, upper = NA, mean = 0, var = 0)
 
       # arrival time for subjects already enrolled before data cut
       dfa1 <- df %>%
         dplyr::arrange(.data$randdt) %>%
         dplyr::mutate(t = as.numeric(.data$randdt - trialsdt + 1),
                       n = dplyr::row_number()) %>%
-        dplyr::mutate(lower = NA, upper = NA, mean = .data$n, var = 0) %>%
+        dplyr::mutate(pilevel = pilevel, lower = NA, upper = NA,
+                      mean = .data$n, var = 0) %>%
         dplyr::bind_rows(df0) %>%
-        dplyr::bind_rows(dplyr::tibble(t = t0, n = n0, lower = NA,
-                                       upper = NA, mean = n0, var= 0)) %>%
-        dplyr::select(.data$t, .data$n, .data$lower, .data$upper,
-                      .data$mean, .data$var) %>%
+        dplyr::bind_rows(dplyr::tibble(t = t0, n = n0, pilevel = pilevel,
+                                       lower = NA, upper = NA,
+                                       mean = n0, var= 0)) %>%
+        dplyr::select(.data$t, .data$n, .data$pilevel, .data$lower,
+                      .data$upper, .data$mean, .data$var) %>%
         dplyr::group_by(.data$t) %>%
         dplyr::slice(dplyr::n()) %>%
         dplyr::ungroup()
@@ -462,7 +459,7 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
       # add predicted from data cut to specified years after data cut
       dfb1t0 <- dfa1 %>%
         dplyr::slice(dplyr::n()) %>%
-        dplyr::mutate(lower = .data$n, upper = .data$n)
+        dplyr::mutate(pilevel = pilevel, lower = .data$n, upper = .data$n)
       dfb1t1 <- dfb1t0 %>%
         dplyr::mutate(t = t1)
 
@@ -527,6 +524,7 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
                          by = c("treatment", "treatment_description")) %>%
         dplyr::mutate(nenrolled = .data$nenrolled + .data$n0) %>%
         dplyr::summarise(n = quantile(.data$nenrolled, probs = 0.5),
+                         pilevel = pilevel,
                          lower = quantile(.data$nenrolled, probs = plower),
                          upper = quantile(.data$nenrolled, probs = pupper),
                          mean = mean(.data$nenrolled),
@@ -540,7 +538,8 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
       # day 1
       df0 <- sum_by_trt %>%
         dplyr::select(.data$treatment, .data$treatment_description) %>%
-        dplyr::mutate(t = 1, n = 0, lower = NA, upper = NA, mean = 0, var = 0)
+        dplyr::mutate(t = 1, n = 0, pilevel = pilevel, lower = NA,
+                      upper = NA, mean = 0, var = 0)
 
       # arrival time for subjects already enrolled before data cut
       dfa1 <- df2 %>%
@@ -548,14 +547,16 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
         dplyr::arrange(.data$randdt) %>%
         dplyr::mutate(t = as.numeric(.data$randdt - trialsdt + 1),
                       n = dplyr::row_number()) %>%
-        dplyr::mutate(lower = NA, upper = NA, mean = .data$n, var = 0) %>%
+        dplyr::mutate(pilevel = pilevel, lower = NA, upper = NA,
+                      mean = .data$n, var = 0) %>%
         dplyr::bind_rows(df0) %>%
         dplyr::bind_rows(sum_by_trt %>%
-                           dplyr::mutate(t = t0, n = n0, lower = NA,
-                                         upper = NA, mean = n0, var = 0)) %>%
+                           dplyr::mutate(t = t0, n = n0, pilevel = pilevel,
+                                         lower = NA, upper = NA,
+                                         mean = n0, var = 0)) %>%
         dplyr::select(.data$treatment, .data$treatment_description,
-                      .data$t, .data$n, .data$lower, .data$upper,
-                      .data$mean, .data$var) %>%
+                      .data$t, .data$n, .data$pilevel, .data$lower,
+                      .data$upper, .data$mean, .data$var) %>%
         dplyr::group_by(.data$treatment, .data$treatment_description,
                         .data$t) %>%
         dplyr::slice(dplyr::n()) %>%
@@ -567,7 +568,7 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
       # add predicted from data cut to specified years after data cut
       dfb1t0 <- dfa1 %>%
         dplyr::slice(dplyr::n()) %>%
-        dplyr::mutate(lower = .data$n, upper = .data$n)
+        dplyr::mutate(pilevel = pilevel, lower = .data$n, upper = .data$n)
       dfb1t1 <- dfb1t0 %>%
         dplyr::mutate(t = t1)
 
@@ -1001,6 +1002,7 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
     # predicted number of events after data cut
     dfb2 = df1 %>%
       dplyr::summarise(n = quantile(.data$nevents, probs = 0.5),
+                       pilevel = pilevel,
                        lower = quantile(.data$nevents, probs = plower),
                        upper = quantile(.data$nevents, probs = pupper),
                        mean = mean(.data$nevents),
@@ -1009,6 +1011,7 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
     # predicted number of dropouts after data cut
     dfb3 = df1 %>%
       dplyr::summarise(n = quantile(.data$ndropouts, probs = 0.5),
+                       pilevel = pilevel,
                        lower = quantile(.data$ndropouts, probs = plower),
                        upper = quantile(.data$ndropouts, probs = pupper),
                        mean = mean(.data$ndropouts),
@@ -1017,6 +1020,7 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
     # predicted number of subjects at risk after data cut
     dfb4 = df1 %>%
       dplyr::summarise(n = quantile(.data$nongoing, probs = 0.5),
+                       pilevel = pilevel,
                        lower = quantile(.data$nongoing, probs = plower),
                        upper = quantile(.data$nongoing, probs = pupper),
                        mean = mean(.data$nongoing),
@@ -1024,16 +1028,17 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
 
     if (!is.null(df)) {
       # day 1
-      df0 <- dplyr::tibble(t = 1, n = 0, lower = NA, upper = NA,
-                           mean = 0, var = 0)
+      df0 <- dplyr::tibble(t = 1, n = 0, pilevel = pilevel,
+                           lower = NA, upper = NA, mean = 0, var = 0)
 
       # observed number of events before data cut
       dfa2 <- df %>%
         dplyr::arrange(.data$totalTime) %>%
         dplyr::mutate(t = .data$totalTime, n = cumsum(.data$event)) %>%
-        dplyr::mutate(lower = NA, upper = NA, mean = .data$n, var = 0) %>%
-        dplyr::select(.data$t, .data$n, .data$lower, .data$upper,
-                      .data$mean, .data$var) %>%
+        dplyr::mutate(pilevel = pilevel, lower = NA, upper = NA,
+                      mean = .data$n, var = 0) %>%
+        dplyr::select(.data$t, .data$n, .data$pilevel, .data$lower,
+                      .data$upper, .data$mean, .data$var) %>%
         dplyr::bind_rows(df0) %>%
         dplyr::filter(.data$t <= tp) %>%
         dplyr::arrange(.data$t)
@@ -1042,9 +1047,10 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
       dfa3 <- df %>%
         dplyr::arrange(.data$totalTime) %>%
         dplyr::mutate(t = .data$totalTime, n = cumsum(.data$dropout)) %>%
-        dplyr::mutate(lower = NA, upper = NA, mean = .data$n, var = 0) %>%
-        dplyr::select(.data$t, .data$n, .data$lower, .data$upper,
-                      .data$mean, .data$var) %>%
+        dplyr::mutate(pilevel = pilevel, lower = NA, upper = NA,
+                      mean = .data$n, var = 0) %>%
+        dplyr::select(.data$t, .data$n, .data$pilevel, .data$lower,
+                      .data$upper, .data$mean, .data$var) %>%
         dplyr::bind_rows(df0) %>%
         dplyr::filter(.data$t <= tp) %>%
         dplyr::arrange(.data$t)
@@ -1059,9 +1065,10 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
                                       (.data$event == 0 &
                                          .data$dropout == 0))),
                          .groups = "drop_last") %>%
-        dplyr::mutate(lower = NA, upper = NA, mean = .data$n, var = 0) %>%
-        dplyr::select(.data$t, .data$n, .data$lower, .data$upper,
-                      .data$mean, .data$var) %>%
+        dplyr::mutate(pilevel = pilevel, lower = NA, upper = NA,
+                      mean = .data$n, var = 0) %>%
+        dplyr::select(.data$t, .data$n, .data$pilevel, .data$lower,
+                      .data$upper, .data$mean, .data$var) %>%
         dplyr::bind_rows(df0) %>%
         dplyr::filter(.data$t <= tp) %>%
         dplyr::arrange(.data$t)
@@ -1254,6 +1261,7 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
     # predicted number of events after data cut
     dfb2 = df1 %>%
       dplyr::summarise(n = quantile(.data$nevents, probs = 0.5),
+                       pilevel = pilevel,
                        lower = quantile(.data$nevents, probs = plower),
                        upper = quantile(.data$nevents, probs = pupper),
                        mean = mean(.data$nevents),
@@ -1263,6 +1271,7 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
     # predicted number of dropouts after data cut
     dfb3 = df1 %>%
       dplyr::summarise(n = quantile(.data$ndropouts, probs = 0.5),
+                       pilevel = pilevel,
                        lower = quantile(.data$ndropouts, probs = plower),
                        upper = quantile(.data$ndropouts, probs = pupper),
                        mean = mean(.data$ndropouts),
@@ -1272,6 +1281,7 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
     # predicted number of subjects at risk after data cut
     dfb4 = df1 %>%
       dplyr::summarise(n = quantile(.data$nongoing, probs = 0.5),
+                       pilevel = pilevel,
                        lower = quantile(.data$nongoing, probs = plower),
                        upper = quantile(.data$nongoing, probs = pupper),
                        mean = mean(.data$nongoing),
@@ -1283,7 +1293,8 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
       # day 1
       df0 <- sum_by_trt %>%
         dplyr::select(.data$treatment, .data$treatment_description) %>%
-        dplyr::mutate(t = 1, n = 0, lower = NA, upper = NA, mean = 0, var = 0)
+        dplyr::mutate(t = 1, n = 0, pilevel = pilevel,
+                      lower = NA, upper = NA, mean = 0, var = 0)
 
 
       # observed number of events before data cut
@@ -1291,10 +1302,11 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
         dplyr::group_by(.data$treatment, .data$treatment_description) %>%
         dplyr::arrange(.data$totalTime) %>%
         dplyr::mutate(t = .data$totalTime, n = cumsum(.data$event)) %>%
-        dplyr::mutate(lower = NA, upper = NA, mean = .data$n, var = 0) %>%
+        dplyr::mutate(pilevel = pilevel, lower = NA, upper = NA,
+                      mean = .data$n, var = 0) %>%
         dplyr::select(.data$treatment, .data$treatment_description,
-                      .data$t, .data$n, .data$lower, .data$upper,
-                      .data$mean, .data$var) %>%
+                      .data$t, .data$n, .data$pilevel, .data$lower,
+                      .data$upper, .data$mean, .data$var) %>%
         dplyr::bind_rows(df0) %>%
         dplyr::filter(.data$t <= tp) %>%
         dplyr::arrange(.data$treatment, .data$treatment_description, .data$t)
@@ -1304,10 +1316,11 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
         dplyr::group_by(.data$treatment, .data$treatment_description) %>%
         dplyr::arrange(.data$totalTime) %>%
         dplyr::mutate(t = .data$totalTime, n = cumsum(.data$dropout)) %>%
-        dplyr::mutate(lower = NA, upper = NA, mean = .data$n, var = 0) %>%
+        dplyr::mutate(pilevel = pilevel, lower = NA, upper = NA,
+                      mean = .data$n, var = 0) %>%
         dplyr::select(.data$treatment, .data$treatment_description,
-                      .data$t, .data$n, .data$lower, .data$upper,
-                      .data$mean, .data$var) %>%
+                      .data$t, .data$n, .data$pievel, .data$lower,
+                      .data$upper, .data$mean, .data$var) %>%
         dplyr::bind_rows(df0) %>%
         dplyr::filter(.data$t <= tp) %>%
         dplyr::arrange(.data$treatment, .data$treatment_description, .data$t)
@@ -1323,10 +1336,11 @@ predictEvent <- function(df = NULL, target_d, newSubjects = NULL,
                     (.data$totalTime > .data$t |
                        (.data$event == 0 & .data$dropout == 0))),
           .groups = "drop_last") %>%
-        dplyr::mutate(lower = NA, upper = NA, mean = .data$n, var = 0) %>%
+        dplyr::mutate(pilevel = pilevel, lower = NA, upper = NA,
+                      mean = .data$n, var = 0) %>%
         dplyr::select(.data$treatment, .data$treatment_description,
-                      .data$t, .data$n, .data$lower, .data$upper,
-                      .data$mean, .data$var) %>%
+                      .data$t, .data$n, .data$pilevel, .data$lower,
+                      .data$upper, .data$mean, .data$var) %>%
         dplyr::bind_rows(df0) %>%
         dplyr::filter(.data$t <= tp) %>%
         dplyr::arrange(.data$treatment, .data$treatment_description, .data$t)